+++ ggml/src/ggml-gemmini/bench_tensor/gemmini_bench_tensor.tpp
@@ -2,6 +2,7 @@
 #include "gemmini_bench_tensor.h"
 #include "../ggml-gemmini-util.h"
 #include <memory>
+#include <cstring>
 
 namespace zerogod
 {
